Improvement for website list
I suggest that we avoid displaying the outdated and obvious news about the MSP2 Nebula client crash, which occurs approximately 95% of the time due to our company's incompetency. Instead, it would be more beneficial to showcase a useful and entertaining GIF on our TV screen that adds value to our business. In recognition of our Head of Safety's diligent efforts, I recommend featuring a humorous GIF that would lighten the mood and provide a pleasant moment for the team. This would not only entertain our employees but also inspire them to appreciate and recognize their coworkers' hard work and accomplishments.